requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
pymongo==4.2.0
Werkzeug==2.2.2
pyqrcode
dnspython
pypng
gunicorn
